requirements.txt
bcrypt==3.2.0
boto3==1.21.4
botocore==1.24.5
cachelib==0.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
cryptography==36.0.1
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.3
Flask-Login==0.5.0
Flask-Session==0.3.2
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
msal==1.17.0
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
s3transfer==0.5.1
six==1.16.0
subnet-utils==0.5.1
urllib3==1.26.8
-e git+https://github.com/Inner-daydream/vpn-enrollment.git@7e57302c3a0e3f75a4043c9a62d42bf6197f5b49#egg=VPNEnrollment
Werkzeug==2.0.3
wireguard==0.2.5
WTForms==3.0.1
